requirements.txt
boto3==1.41.5
botocore==1.41.5
ensure==1.0.2
fastapi==0.128.0
joblib==1.5.2
numpy==1.23.5
pandas==2.3.3
pendulum==3.1.0
pydantic==2.12.5
python-dotenv==1.2.1
python_box==6.0.2
pyvi==0.1.1
PyYAML==6.0.3
Requests==2.32.5
runpod==1.8.1
scikit_learn==1.8.0
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
uvicorn==0.40.0
